% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-products.R
\name{available_products}
\alias{available_products}
\title{Retrieve a list of tracked products}
\usage{
available_products()
}
\value{
character vector
}
\description{
Each product name can be used for the other API endpoints.
}
\examples{
available_products()
}
