% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaked-package.R
\docType{package}
\name{croaked}
\alias{croaked}
\alias{croaked-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/croaked}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/croaked/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
