% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-cycle.R
\name{cycle_details}
\alias{cycle_details}
\title{Single cycle details}
\usage{
cycle_details(product, cycle)
}
\arguments{
\item{product}{length 1 character vector product name (e.g. \code{python})}

\item{cycle}{length 1 character vector with product cycle (e.g. \code{3.9})}
}
\value{
data frame
}
\description{
Get details of a single cycle
}
\examples{
cycle_details("python", "3.9")
}
