% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaked-package.R
\docType{package}
\name{croaked}
\alias{croaked}
\alias{croaked-package}
\title{Tools to Query the endoflife.date Service}
\description{
The endoflife.date site is a community-maintained project to document
end-of-life dates, and support lifecycles of various products. Tools are provided to
query this end-of-life metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/croaked}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/croaked/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
